Merge develop in dev/cmd-evaluate

+++ proksee/assembly_database.py
@@ -144,13 +144,13 @@ class AssemblyDatabase:
         Returns the N50 quantile for the specified species and quantile value.
 
         RETURNS
-            quantile (int): the N50 quantile for the specified species and quantile value
+            quantile (float): the N50 quantile for the specified species and quantile value
                 returns None if the species on quantile is missing
         """
 
         if species in self.database:
             n50_quantiles = self.database[species][self.N50_QUANTILES]
-            quantile = int(n50_quantiles[value]) if value in n50_quantiles else None
+            quantile = float(n50_quantiles[value]) if value in n50_quantiles else None
 
         else:
             quantile = None
@@ -162,13 +162,13 @@ class AssemblyDatabase:
         Returns the number of contigs quantile for the specified species and quantile value.
 
         RETURNS
-            quantile (int): the number of contigs quantile for the specified species and quantile value
+            quantile (float): the number of contigs quantile for the specified species and quantile value
                 returns None if the species on quantile is missing
         """
 
         if species in self.database:
             contigs_quantiles = self.database[species][self.CONTIGS_QUANTILES]
-            quantile = int(contigs_quantiles[value]) if value in contigs_quantiles else None
+            quantile = float(contigs_quantiles[value]) if value in contigs_quantiles else None
 
         else:
             quantile = None
@@ -180,13 +180,13 @@ class AssemblyDatabase:
         Returns the L50 quantile for the specified species and quantile value.
 
         RETURNS
-            quantile (int): the L50 quantile for the specified species and quantile value
+            quantile (float): the L50 quantile for the specified species and quantile value
                 returns None if the species on quantile is missing
         """
 
         if species in self.database:
             l50_quantiles = self.database[species][self.L50_QUANTILES]
-            quantile = int(l50_quantiles[value]) if value in l50_quantiles else None
+            quantile = float(l50_quantiles[value]) if value in l50_quantiles else None
 
         else:
             quantile = None
@@ -198,13 +198,13 @@ class AssemblyDatabase:
         Returns the assembly length quantile for the specified species and quantile value.
 
         RETURNS
-            quantile (int): the assembly length quantile for the specified species and quantile value
+            quantile (float): the assembly length quantile for the specified species and quantile value
                 returns None if the species on quantile is missing
         """
 
         if species in self.database:
             length_quantiles = self.database[species][self.LENGTH_QUANTILES]
-            quantile = int(length_quantiles[value]) if value in length_quantiles else None
+            quantile = float(length_quantiles[value]) if value in length_quantiles else None
 
         else:
             quantile = None
